pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "mini-agent"
version = "0.1.0"
description = "A teaching project for building production-grade AI agents from scratch"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
dependencies = [
    "openai>=1.0.0",
    "pydantic>=2.0.0",
    "python-dotenv>=1.0.0",
    "rich>=13.0.0",
    "tiktoken>=0.5.0",
]

[project.scripts]
mini-agent = "mini_agent.main:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["mini_agent*"]
